Fix 64-bit compatibility for Dive_Logger, Divelogger2, H20FlowInc, SCUBA_Dive_Logging, Snail_Mail, Thermal_Controller_v3
- Dive_Logger, Divelogger2, SCUBA_Dive_Logging: Fix malloc to use sizeof() instead
  of hardcoded 4-byte values for header size and alignment, so it works correctly
  on 64-bit where pointers and size_t are 8 bytes
- H20FlowInc: Fix POLANDBALL malloc minimum size class from 16 to 32 bytes; on
  64-bit struct list_node is 16 bytes so a free chunk needs at least 32 bytes
  (8 header + 16 list_node + 8 footer) to avoid overflowing into adjacent chunks
- SCUBA_Dive_Logging: Fix time_t and datetime_type to use uint32_t instead of
  unsigned long so wire format is consistent between 32-bit and 64-bit
- Snail_Mail: Fix cgc_rand() call (was incorrectly calling rand() without cgc_ prefix)
- Thermal_Controller_v3: Fix malloc block size calculation to use sizeof(cgc_size_t)
  instead of hardcoded 4 so freelist blocks are correctly sized on 64-bit

+++ challenges/SCUBA_Dive_Logging/include/cgc_dates.h
@@ -27,7 +27,9 @@ THE SOFTWARE.
 #ifndef DATES_H
 #define DATES_H
 
-typedef unsigned long time_t;
+#include "cgc_stdint.h"
+
+typedef uint32_t time_t;
 
 typedef struct {
